requirements.txt
django==2.0.4
gunicorn
whitenoise
django-colorfield

psycopg2
psycopg2-binary
dj-database-url
